requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
google-api-core
google-generativeai
geopandas
networkx